systemtests: pull snapshot images once
When creating Containers, if the image contains the String 'latest', and
it is the first time we are creating a Container for that image in this
process, then we set imagePullPolicy to Always else IfNotPresent

Why:
This is to help when running in a local environment. Ideally minikube would
be clean before running, but this is not automated. A stale testclients
image could cause the tests to fail. This code helps by pulling snapshot
images the first time a container is created.

Signed-off-by: Robert Young <robeyoun@redhat.com>