Update FilterInvoker javadoc
When refactoring invokers we changed the logic so that if any
FilterInvoker indicates it is interested in a message (via
shouldHandle[Request|Response), then all FilterInvokers in the chain are
eligble to handle that message. The FilterInvoker is responsible for
doing a simple forward if it receives a decoded message that it isn't
interested in.

Signed-off-by: Robert Young <robeyoun@redhat.com>