Release script updates semver checks for kroxylicious-api
After a release we should change the reference version to the released
version so that the next SNAPSHOT is compared to the just-released
version.

We may also need to re-enable semver checking if we have disabled it
during development. Our deprecation policy allows us to make breaking
changes across minor versions while we are pre-1.0.0. We still like
having the plugin warn us of inadvertent API changes, so when we do want
to break pubic APIs we switch off semver checks until release time.

Signed-off-by: Robert Young <robeyoun@redhat.com>