Remove ApiVersionsService
Deprecated since 0.3.0 and now eligible for destruction by our policy.

We haven't used it yet and the utility of it to Filters is unclear. The
original rationale was that Filters can hijack the connection and use
sendRequest to send messages to the broker, messages constructed by the
Filter. We need to ensure we only use fields that the broker
understands, so by offering an API to access the supported versions a
Filter could discover what's available and react to that.

For now we will remove the deprecated class and can re-implement
it if anyone needs this functionality. For now the Filters can work
around it by intercepting the ApiVersionsResponse themselves which
should work for the typical case where the Filter wants to do some work
after the api versions are negotiated.

Signed-off-by: Robert Young <robeyoun@redhat.com>