Enable user to configure provider
This will enable the user to specify a provider so that if that provider
was not available, the Filter initialization will explosively fail,
protecting them from unexpectedly loading the transformation from
another provider. This would be useful if they want to use something
like a FIPS certified provider with transformation names that match the
JVM. We want to fail early instead of loading non-FIPS compliant
algorithms.

Signed-off-by: Robert Young <robeyoun@redhat.com>