Fix #1348: Reword the Record Encryption documentation describing the admin role
The intent is to try to convey that the fact that something outside the system (an admin) needs to be responsible for creating KEKs that will be used by Record Encryption.
It needs to provide guidance as to what permissions that user needs.  This should give the reader the information they need to map that responsibilty into their organisational
structure.

Signed-off-by: kwall <kwall@apache.org>