record-encryption: log details about Cipher origin
during the one-time check that the Ciphers are available, we log details
about the algorithm and provider.

Why:
Users may wish to plug in alternate providers for Cipher algorithms, for
example to use a fips-certified bouncycastle implementation instead of a
platform implementation. This can be achieved by configuring the
environment so that the alternative provider is preferred but we want to
expose where the algorithms are loaded from to give users confidence that
they have configured their environment correctly.

Signed-off-by: Robert Young <robeyoun@redhat.com>