Validate the incoming schema global id (if present) against the expected schema id  (#1417)
If the producer is using schemas, the schema global id will be embedded in the record (as a header headers or serialised in the preamble of the key/value body).  With this change, that schema id is validated against the expected global id.  If the two don't match, the record will be rejected.

If the producer is not using schemas, there is no change in behaviour.

Signed-off-by: kwall <kwall@apache.org>