Exclude transitive dependency on lombok
From the Lombok docs https://projectlombok.org/setup/maven:
> To set up lombok with any build tool, you have to specify that the lombok dependency
is required to compile your source code, but does not need to be present when
running/testing/jarring/otherwise deploying your code. Generally this is called
a 'provided' dependency.

We can remove this, less jars to worry about on the classpath

Signed-off-by: Robert Young <robeyoun@redhat.com>