Setup Sonarcloud connected mode (IDE integration for Intellij)
The Sonar plugin for IntelliJ has the ability to use rules from a project definition in Sonarcloud.
This should help developers be more productivity as the feedback they get in their IDE should
match the rules imposed during the CI workflow.  Hopefully this means fewer iterations before a PR is good to merge.

https://docs.sonarsource.com/sonarlint/intellij/team-features/connected-mode-setup/#bind-using-shared-configuration

Signed-off-by: kwall <kwall@apache.org>