A minimal Kubernetes operator for the proxy (#1560)
This is based on the existing kubernetes-examples/network-topologies/portperbroker_plain mainfests.
It's also based on the JOSDK's getting started, applying the same patterns.
One point of difference with JOSDK is this uses a CRD-first approach, generating the Java CR classes using Fabric8's maven plugin. The CRD in `META-INF/fabric8` is the SoT for now. See https://github.com/operator-framework/java-operator-sdk/issues/2561
That works because JOSDK uses Fabric8's Kubernetes client.
One implication of that is we end up with classes which depend on the CRD schema version.
So when we eventually evolve our CRD schema (e.g. add a v1beta1) we'd end up with a whole other class hierarchy.

Signed-off-by: Tom Bentley <tbentley@redhat.com>