operator: add part-of and managed-by labels
Adding a labelSelector to the KubernetesDependent annotation means that
when setting up the watches via fabric8 it can pass down the label
selector. This means the API server won't send events to the operator
for all Deployments/Services/Secrets, only those managed by the
operator.

Part-of is just one of the recommended labels, it will allow users to
use label selectors to locate the resources associated with a particular
proxy CR

https://kubernetes.io/docs/concepts/overview/working-with-objects/common-labels/

Signed-off-by: Robert Young <robeyoun@redhat.com>