Enable users to supply labels to Kroxylicious podTemplate
This adds a podTemplate property to KafkaProxy that is mapped to
a io.fabric8.kubernetes.api.model.PodTemplateSpec. This spec is
used as the basis for the Deployment podTemplate.

We want to be careful about which properties we enable users to work
with as, once released, we will be on the hook for compatibility. So for
a start we only tolerate passing through metadata.labels. In future we
will like add further elements of the PodTemplateSpec like resource
limits/requests.

Signed-off-by: Robert Young <robeyoun@redhat.com>