operator: Use absolute service host for brokerAddressPattern
Enables client pods in arbitrary namespaces to connect to the proxy via
the service created by the operator.

The brokerAddressPattern was previously the service name, which is
resolved relative to the client pod namespace. So after clients
retrieved cluster metadata via the proxy they could not connect to the
brokers as they attempted to resolve the broker relative to their own
namespace.

ie bootstrapping with returned my-cluster.my-proxy.svc.cluster.local:9292
returned node addresses like my-cluster:${brokerPort} and the
resolution failed.

Signed-off-by: Robert Young <robeyoun@redhat.com>