fix: proxy can't handle ApiVersion version from client ahead of proxy
If we receive an ApiVersion request at a version higher than the proxy
supports we want to tell the client that it is unsupported along with
the highest ApiVersions version that is supported by broker and proxy.

To do this the proxy intercepts the ApiVersions request, substituting
it's own v0 request to the broker to obtain the broker supported
versions. Then when the proxy intercepts the response it sets the
UNSUPPORTED_VERSION error on the response to tell the client to try
again with the highest supported ApiVersions version.

Signed-off-by: Robert Young <robeyoun@redhat.com>