Require opt-in to allow testing configuration
Adds a concept of configuration features. We want to add some features to the
proxy to assist in testing that we want to be disallowed unless the user
has opted into them, telling the proxy that these features are unlocked.

By default the proxy has no features enabled and will log an error and
fail to start if any testing configurations are specified in the config
file.

To opt in to using test-only configuration the user must supply either:
1. a system prop `-DKROXYLICIOUS_UNLOCK_TEST_ONLY_CONFIGURATION=true`
2. an env var `KROXYLICIOUS_UNLOCK_TEST_ONLY_CONFIGURATION=true`
with the system-prop taking precedence.

Integration tests behave the same way,you have to enable the feature
explicitly in your test to use any test-only features.

Why:
We plan to add some opt-in configuration to enable overriding the maximum
version the proxy accepts for an ApiKey. This is test only functionality
and we want to ensure this isn't accidentally applied in production
without some kind of opt-in hoops to leap through.

Signed-off-by: Robert Young <robeyoun@redhat.com>