Fix #1436 : remove forwardPartialRequests feature of record validation
Why: in the commentary of #1436 it was concluded that the inclusion of forwardPartialRequests option in the RecordValidation filter was an error. The problem
being that if a client is sending incorrect data, then stopping immediately is the safe course of action.   Accepting the correct data is likely to lead to
more harm than good.   We said we wanted to drop the feature (without deprecation).

Signed-off-by: kwall <kwall@apache.org>