Fix race condition at DEK exhaustion time
It was possible for a thread to successfully be allocated the desired
number of encryptions, then for it to discover that another thread had
destroyed the DEK for encryption (due to incrementing the atomiclong
before trying to atomically obtain a cryptor reference). Eg:

thread 1: obtains 10 encryptions
thread 2: fails to obtain 10 encryptions, destroys DEK for encryption
thread 1: tries to bump encryptor reference count, fails

With this change we acquire a Cryptor reference count before we attempt
to allocate it the desired amount of encryptions. Then if we cannot
obtain the required amount of encryptions we release the reference.
This means the final allocation of encryptions will not be wasted. The
flow becomes:

thread 1: bumps encryptor reference count
thread 2: bumps encryptor reference count
thread 1: obtains 10 encryptions
thread 2: fails to obtain 10 encryptions, destroys DEK for encryption
thread 2: releases reference
thread 1: carries on merrily encrypting, eventually releasing reference

Signed-off-by: Robert Young <robeyoun@redhat.com>