Prevent invalidation of unexpected encryption DEKs
If a thread finds that a DEK obtained from the cache has been destroyed
or exhausted, it initiates a cache invalidation then a retry. This
should then load a new DEK and carry on happily. The issue is that under
heavy load multiple threads could initiate the cache invalidation. One
of the invalidations will remove the DEK and the other may invalidate
the next DEK, or have undefined results if an async load is in flight.

To solve this we use computeIfAbsent to only remove the entry from the
cache if the value is the invalid DEK.

Signed-off-by: Robert Young <robeyoun@redhat.com>