build-image.sh: enable Containerfile selection
Set env var CONTAINERFILE=my-alternative-Dockerfile to select an
alternative Containerfile to build (relative to the root directory
of the project). The default is "Dockerfile".

With the addition of the operator Docker image it is helpful to be able
to use the existing script to build it and push to quay.

To build the operator you can run:
```
REGISTRY_DESTINATION=quay.io/my-quay-username/operator CONTAINERFILE=Dockerfile.operator ./scripts/build-image.sh
```

Signed-off-by: Robert Young <robeyoun@redhat.com>