operator: expose cluster with in-cluster-only SNI
A KafkaProxy must declare a single listener specifying a port
A KafkaProxy cluster must declare a single hostname. The hostname
cannot be a wildcard hostname like *.xyz.com. The hostname must
contain a '%' hostname section, this cannot be the first character.
The hostname must end with <ns>.svc.cluster.local, where <ns> is
the namespace of the KafkaProxy.

The cluster must declare a non-empty advertisedBrokers array

The listener must also define a single certificateRef. That
certificateRef must be to a secret containing two keys:
1. `tls.crt` containing the server certificate in x509 format
2. `tls.key` containing the server private key in PKCS8 format
following the naming conventions of cert-manager.

If the hostname is declared is 'my-cluster-%.my-na.svc.cluster.local`
and the listener port is 9092 and the advertiserBrokers are 0,1 and 2.
Then the operator will manifest kubernetes Services:
1. my-cluster-bootstrap, 9092 -> 9092
2. my-cluster-broker-0, 9092 -> 9092
3. my-cluster-broker-1, 9092 -> 9092
4. my-cluster-broker-2, 9092 -> 9092

The VirtualCluster on the proxy will be configured to expect SNI
on 9092 with bootstrap `my-cluster-bootstrap.my-ns.svc.cluster.local`
and brokerpattern `my-cluster-broker-$(nodeId).my-ns`.

The VirtualCluster on the manifested proxy will also be configured to
mount the secret at
`/opt/kroxylicious/secrets/tls/<listenerName>/<certificateSecretName>`
and the proxy configuration will load a KeyPair from that location.

Put it all together and clients in the cluster can bootstrap with
`my-cluster-bootstrap.my-ns.svc.cluster.local:9092`.

Signed-off-by: Robert Young <robeyoun@redhat.com>