Add support for a KeyPairSet for downstream tls keys
Users can configure a list of KeyPairs for a virtual clusters server
TLS. When selecting the KeyPair for that virtual cluster we select one
of the keypairs. The criteria for selection is:

1. The certificate of the keypair has SANs containing the virtual
   cluster bootstrap hostname.
2. Or, if the cert has no SAN, the bootstrap address matches the subject
   common name.

It is an error to match multiple certificates.

Why:

We are investigating implementing the Gateway API. In the Gateway API a
Gateway has listeners, equivalent to an ip+port. A listener can have an
associated list of certificateRefs. But in the Proxy currently multiple
virtual clusters attached to the same port may have distinct
certificates, as they are exposing different hostnames. The Gateway API
doesn't give us a way to associate certain certificates with certain
routes.

So this is one idea to work around this. We would instead supply the
virtual cluster with all the certificates of the gateway, and then
select the appropriate certificate for that Virtual Cluster by looking
at each certificate's SANs. This demonstration expects the exposed
bootstrap hostname to match a SAN in exactly one of the certificates.

Note that for SNI we will not currently be able to check that all the
broker ids have an appropriate SAN (once templated into the brokerPattern)
as we are not aware of the node ids of the target cluster in advance.
We can only check our bootstrap hostname.

Signed-off-by: Robert Young <robeyoun@redhat.com>