record encryption: add UnresolvedKeyPolicy
Adds a Record Encryption configuration property `unresolvedKeyPolicy`
with possible values:
- PASSTHROUGH_UNENCRYPTED (default)
- REJECT

If we cannot resolve a Key for some topics in a produce request then we
will decide what to do based on this policy.

PASSTHROUGH_UNENCRYPTED will not modify the data for those unresolved
topics and pass them along the chain. This is the current behaviour.

REJECT will cause the entire produce request to be rejected if any topic
could not be resolved to a Key, an error response will be sent to the client.

Example configuration:
```
filterDefinitions:
- name: encrypt
  type: RecordEncryption
  config:
    kms: VaultKmsService
    kmsConfig:
      vaultTransitEngineUrl: http://vault:8200/v1/transit
      vaultToken:
        password: token
    selector: TemplateKekSelector
    selectorConfig:
      template: "KEK_${topicName}"
    unresolvedKeyPolicy: REJECT
```

Why:
Currently when we are unable to resolve a Key for a topic, we interpret
that as meaning the topic should not be encrypted. This fail-open policy leaves the
possibility for misconfigurations to permissively allow unencrypted data
to flow into the target cluster.

Some users may wish to set up the proxy such that only encrypted traffic
should flow through it, in this case they can have a harsher fail-closed
policy in the case where we can't resolve a key for a topic. This will
be safer, in terms of PII/data-governance requirements as the proxy will
never decide to passthrough unencrypted data. Any misconfiguration in
the Proxy or KMS where the naming doesn't align will result in failure
to produce, rather than unencrypted data being sent to the target
cluster.

Signed-off-by: Robert Young <robeyoun@redhat.com>