record encryption: add defaultTopicEncryptionPolicy
If defaultTopicEncryptionPolicy is set to REQUIRE_ENCRYPTION then we
require all topics in a ProduceRequest to have their data encrypted. If
we fail to resolve an alias for any topic in the request then we reject
the entire produce request and return an InvalidRecordException to the
client.

If defaultTopicEncryptionPolicy is not set then we use the legacy
behaviour, using the KMS response to determine whether we should encrypt
or passthrough a topic's data unencrypted.

Why:

Currently we are using a problematic strategy where we use the response
from the KMS to determine whether or not to encrypt. If we could not
resolve a key for a topic, we pass it through unencrypted. This
permissive strategy means a misconfiguration could allow unencrypted
data through due to a misconfiguration in the KMS or proxy.

Some users would prefer a harsher default policy where we require ALL data sent
to the proxy to be encrypted. This means they will not be able to
produce to unencrypted topics, but would have some protection against
unencrypted data being written to their target cluster unexpectedly.

In future we could imagine having different policies per topic, which is
why we have named this the default topic encryption policy. In the
absence of any topic specific policy this is the applied policy. This
would enable users to safely use a virtual cluster for encrypted and
unencrypted data as we can determine if a topic should be encrypted,
rather than using the KMS response.

Signed-off-by: Robert Young <robeyoun@redhat.com>