Enable Filters to obtain topic names for topic ids
Adds CompletionStage<Map<Uuid, String>> getTopicNames(Set<Uuid> topicUuids)
to FilterContext. This returns a stage that will be completed with a
Map containing an entry for every topicUuid requested, if any can not
be mapped to a name the stage is completed exceptionally.

The Map returned is allowed to contain mappings for a superset of uuids.
This is to enable us to return an unmodifiable wrapper around a single
ConcurrentMap stored on the VirtualClusterModel. The assumption is that a
topic id will always refer to a single topic name forever.

If the proxy has not encountered those topicIds before, we send an out
of band v12 Metadata request to the broker to obtain the topic names.
The Proxy intercepts the response and learns the name mappings. We can
then complete the future with the Map.

If the proxy has already learned those mappings, then we can return a
completed stage immediately.

Why:

KIP-516 introduced Topic Ids to address some issues with using topic
names as the identifier. Stale topics (topic named X is deleted and a
new topic named X is created) cause edge cases around log dir cleanup.

Many RPCs now operate on topic ids instead of names. This poses a
challenge to some of our Filters. For example users may wish to define
policies for their topics like "encrypt all data for topic named X,Y or
Z". Generally, users will be working with topic names and Kafka's
internal topic ids are an implementation detail they are unlikely to be
interested in.

This means eventually we are going to have to handle ProduceRequests
identifying topics by topic id, mapping them to topic names so that we
can apply the users key selection policy.

More generally there is a class of Filters that are topic-oriented,
users will want different behaviour per-topic. Mask this topic, schema
validate this topic, etc. We should enable users to declare their intent
using topic names and give Filter Authors the means to obtain the filter
names.

Signed-off-by: Robert Young <robeyoun@redhat.com>