Add failing integration test for name lookup issue
Currently the name lookup mechanism breaks Filter composability because
it exposes the real topic names from the backing cluster. Some other
Filters in the chain may be manipulating topic names and wish to apply
equivalient manipulations to the topic names retrieved from the target
cluster.

The integration test shows that even though we have a multi-tenant
filter between the topic name stamping filter and the target broker, the
stamping filter is still able to access the names from the target
cluster.

We likely need some mechanism to let every Filter in the chain intercept
the topic id -> name data.

Signed-off-by: Robert Young <robeyoun@redhat.com>