bugfix: service did not select pods of a certain KafkaProxy
The service selector just targeted any pod with app=kroxylicious label,
which is true for all KafkaProxy pods. We need the service to only
target the pods of a specific KafkaProxy.

To do this, I've included the standard labels on the pod (which includes
the proxy name), and made the service select all the pod labels.
Similarly the deployment selector is updated to target all labels of the
pods.

Signed-off-by: Robert Young <robeyoun@redhat.com>