Add wait for deployment ready
There appears to be some test instability caused by the first download
of the operator image. This takes ~70seconds on my machine to pull from
quay. So some assertions that waited for Services to enter particular
states were exceeding their 10 second wait.

We need a big wait for the initial image pull. Ideally we could directly
pull it in the test setup.

Signed-off-by: Robert Young <robeyoun@redhat.com>