Testcontainers uses docker mirror
Dockerhub rate limits kill our integration tests. Lets default to using
a mirror with exceptions for registries that we know are unlimited.

Note that our failure was on a registry docker.redpanda.com, which
looked like it was eventually backed by dockerhub, so we are
pessimistically defaulting everything to the mirror. This may bite us if
we want to depend on dockerhub images that aren't mirrored, we'll see
how it goes.

Signed-off-by: Robert Young <robeyoun@redhat.com>