Extract KafkaProxyIngress CRD
This part of the Operator design introduces the concept of a
KafkaProxyIngress.

* A KafkaProxyIngress is associated with a single KafkaProxy.
* A VirtualKafkaCluster must reference at least one ingress of the proxy
  it references via proxyRef
* Multiple VirtualKafkaCluster's may reference the same
  KafkaProxyIngress.

The first ingress type implemented is clusterIP. A clusterIP ingress for
a virtual cluster will:
1. create a single Service for that ingress+cluster combination
2. map the ports of the Service to the relevant ports on the pod
3. create a VirtualCluster configuration in the Proxy Config, using
   PortPerBroker and assigning it a range of ports
4. Configure the container ports on the deployment

Initially we will only support a single clusterIP ingress per
KafkaProxy. This is because the implementation depends on a port per
broker Gateway on the proxy side, where we can identify the cluster and
node the client wants to connect to from the port. Using a naive port
assignment strategy to allocate contigious ports allows for edge cases
where Services may temporarily point at proxy ports for which the
identify has changed. For example removing a VirtualCluster could cause
another VirtualCluster to be allocated the ports of the old one.
See https://github.com/kroxylicious/kroxylicious/issues/1902

Note that currently we naively pick the first ingress of the first
virtual cluster as the chosen one. This means that if the first virtual
cluster is invalidated for some other reason, like an unresolved
reference, that subsequent ingresses are still considered in conflict.

Signed-off-by: Robert Young <robeyoun@redhat.com>