refactor: deterministically order labels
Technically the labels don't need to be ordered, it doesn't matter to
kubernetes if the order changes. But in our DerivedResourcesTest we
convert to YAML when the reconciled resources are not equal to our
expectation, and then compare the YAML strings to produce a more useful
output for humans.

These strings currently will often have labels reported as mismatched,
which is highlighted by the IDE and distracts from the other real
differences between the resources that caused the failure.

Signed-off-by: Robert Young <robeyoun@redhat.com>