refactor: introduce proxy model and early dependency resolution
1. Resolves all dependencies of a KafkaProxy up front and record which
virtual clusters have unresolved dependencies in the Context.
2. Adds a ProxyModel to wrap up common code where we resolve
   dependencies and build the ingress model.

We found that calling in and out of the Context to record whether a
VirtualCluster was accepted was brittle within ProxyConfigSecret. It's
better to have it self-contained and not have to remember to recheck the
broken clusters at multiple points in the code.

The resolver is now the single point where we look up secondary
resources from the context for (apart from when reporting cluster
statuses).

Signed-off-by: Robert Young <robeyoun@redhat.com>