Report unresolved KafkaProxyIngress
This introduces a status condition for when a referenced KafkaProxyIngress can
not be found.

We also refactor where this is detected as the code reporting these
statuses was becoming unwieldy. The refactor includes:

1. Resolving all dependencies of a KafkaProxy up front and record which
virtual clusters have unresolved dependencies in the Context.
2. Adding a ProxyModel to wrap up common code where we resolve
   dependencies and build the ingress model.
3. ProxyConfigSecret does not re-check which clusters are broken using
   static methods. Keeping the logic more self-contained.

We found that calling in and out of the Context to record whether a
VirtualCluster was accepted was brittle within ProxyConfigSecret. It's
better to have it self-contained and not have to remember to recheck the
broken clusters at multiple points in the code.

The resolver is now the single point where we look up secondary
resources from the context for (apart from when reporting cluster
statuses).

The ProxyModel aims to contain a logical model of what is going to be
manifested which can be used by all participating reconciler components
to get a shared view of which ingresses are exposed, what ports they are using
etc.

Signed-off-by: Robert Young <robeyoun@redhat.com>