Rename KafkaClusterRef to KafkaService
Why:
KafkaClusterRef's naming causes some pain. Our Custom Resources often
reference other Custom Resources, and our naming convention for these
reference properties is *Ref, like `proxyRef` or `filterRefs`. This
means we would awkwardly have clusterRefRef to reference this CR. Also
with the arrival of LocalRef, Ref now has a stronger meaning in our
codebase that would be confusing. We landed on KafkaService as a better
name for this abstraction.

Signed-off-by: Robert Young <robeyoun@redhat.com>