refactor: replace targetCluster with targetKafkaServiceRef
We've decided that VirtualKafkaCluster should always have a reference to
a KafkaService and it will be the KafkaService that will contain a
bootstrap or a reference to some other CRD like a Strimzi Kafka
resource. This means the targetCluster object isn't required any more.

See: https://github.com/kroxylicious/design/pull/61

Signed-off-by: Robert Young <robeyoun@redhat.com>