Add nodeIdRanges property to KafkaService
The nodeIdRanges are passed through as part of the gateway configuration
for clusterip ingress.

The nodeIdRanges determine how many ports are required by the cluster ip
ingress. We calculate the total number of node ids represented by all
ranges for the ingress. The required port count is this sum plus 1 port
for the bootstrap port.

Signed-off-by: Robert Young <robeyoun@redhat.com>