Add nodeIdRanges property to KafkaService
The nodeIdRanges are passed through as part of the gateway configuration
for clusterip ingress. A nodeIdRange has a start (inclusive) and end
(inclusive) node id along with an optional name. The range end must be
greater than or equal to the start.

The nodeIdRanges determine how many ports are required by the cluster ip
ingress. We calculate the total number of node ids represented by all
ranges for the ingress. The required port count is this sum plus 1 port
for the bootstrap port.

The default name for a range is 'range-${itemIndex}' such that:

```yaml
nodeIdRanges:
  - start: 1
    end: 2
  - start: 3
    end: 4
```

is equivalent to:

```yaml
nodeIdRanges:
  - name: range-0
    start: 1
    end: 2
  - name: range-1
    start: 3
    end: 4
```

Signed-off-by: Robert Young <robeyoun@redhat.com>