Add integration test for CRD validation
Tests that:
1. myriad invalid custom resources will not be accepted by k8s
2. all the in-* files used by our DerivedResourcesTest are valid, this
  is handy because it is a crucial test, but since it is mocked it is
  possible for us to test against different data than would be allowed
  in via the k8s APIs.

Why:
Our CRDs contain a lot of declared constraints. Required fields, types,
length restrictions, patterns. We want some quick feedback that we have
specified things correctly. Checking directly with a json schema
validation library is a bit tricky since k8s uses an odd standard with
their own extensions. Instead we test against minikube to get some real
integration with a k8s API server.

Note that when we assert which error messages were returned from the k8s
APIs we are using contains to check something minimal showing that the
resource was invalidated for the expected reason. The detailed error
message is more likely to change with k8s version, so we don't do an
exact match on the whole error.

Signed-off-by: Robert Young <robeyoun@redhat.com>