refactor: use jackson java8 date module
* Remove unused pattern from Condition#setLastTransitionTime and hoist
  the annotations onto the field. We still need the JsonFormat
  annotation to persuade the module to emit a String, rather then the
  default number format.
* Add the jsr310 jackson module to the operator
* Register the JavaTimeModule in all relevant ObjectMappers. We have the
  default fabric8 ObjectMapper, the proxy config mapper and some test
  usages.

Initially we had some problems with a pattern on the setter causing
Instant to fail at serialization time. This is because it needs a
timezone specified on the ObjectMapper or annotation, or for the
annotation to be omitted, in which case it will use the Instant's
toString for serialization. We removed the pattern so that we use
Instant#toString.

Signed-off-by: Robert Young <robeyoun@redhat.com>