operator: cluster reconciler considers ingress proxyRefs
It will mark the cluster as ResolvedRefs False with a transitive
resolution failure if a referenced ingress references a non-existent
proxy or a proxy that is different to the virtual cluster's proxy.

It might be better if we could disambiguate it, it may be a valid
reference to another proxy, but we mark it as a ResolvedRefs failure
which is misleading.

Signed-off-by: Robert Young <robeyoun@redhat.com>