Rework DependencyResolve class hierarchy
ResolutionResult is really a ProxyResolutionResult, relevant only to a
full-proxy reconciliation.
UnresolvedReferences is now ClusterResolutionResult, it describes the
result of a resolution attempt which could have been successful, or
failed with problems like dangling refs or ResolvedRefs: False
conditions on the dependencies.
Lifted ClusterResolutionResult to the top level as it has a life outside
of a ProxyResolutionResult

Signed-off-by: Robert Young <robeyoun@redhat.com>