Refactor DependencyResolver
The code was a bit backwards in that we treated VKC resolution the same
as a KafkaProxy resolution, then extracted the single result we cared
about from the ProxyResolutionResult. We really want to compose the
KafkaProxy resolution result from multipl VirtualKafkaCluster
resolutions. This Refactor arranges the code that way. This enables us
to make the VirtualKafkaCluster conditions check something specific to
proxy reconciliation instead of passing down a boolean to control the
behaviour.

Signed-off-by: Robert Young <robeyoun@redhat.com>