KafkaProxyReconciler error on stale Referent status
If we encounter a KafkaProxyIngress, KafkaService or KafkaProtocolFilter
with stale status during KafkaProxy reconciliation we:
1. abort the reconcilation
2. log at INFO about the stale referent
3. do not update the KafkaProxy status

When the KafkaProxyReconciler Informer receives a create/update event
for a VirtualKafkaCluster with stale status, we ignore that event and
do not trigger any KafkaProxy reconciliations

Why:
If the metadata.generation of a Referent resource is ahead of the
observed generation in its status we infer that it has not been
reconciled yet and is in an indeterminate state. The
KafkaProxyReconciler should not act on it. Instead we error out on the
assumption the Referent's reconciler will soon apply a new status,
which will in turn trigger a reconciliation of the KafkaProxy.

Signed-off-by: Robert Young <robeyoun@redhat.com>