Test+generate procedure docs from a single source of truth
"Procedures" are a specific type of documentation content
(see https://redhat-documentation.github.io/modular-docs/#con-creating-procedure-modules_writing-mod-docs)
made up of a sequence of steps.

We can generate this kind of documentation from a YAML file which can also be used to execute the steps of the procedure.
Executing the steps of the procedure via junit gives us an automated way of testing that what is documented actually works.

Signed-off-by: Tom Bentley <tbentley@redhat.com>